refactor: separate out validation into its own lib function
